// src/types/types.ts
export interface AppConfig {
  appId?: string
  appName: string
  appNameShort: string
  appStore: string
  configName: string
  darkTheme: Theme
  defaultWallets: string[]
  knowledgeBase: string
  lightTheme: Theme
  notificationServers: string[]
  phoneNumber: string
  referralServers?: string[]
  supportsEdgeLogin: boolean
  supportEmail: string
  supportSite: string
  termsOfServiceSite: string
  website: string
  extraTab?: {
    webviewUrl: string
    tabTitleKey: LocaleStringKey
    extraTabBarIconFont: string
    extraTabBarIconName: string
  }
  extraTab2?: {
    tabTitleKey: LocaleStringKey
    extraTab2BarIconFont: string
    extraTab2BarIconName: string
  }
  hideIoniaRewards?: boolean
}